modify

+++ do_throughput.sh
@@ -17,7 +17,9 @@ case "${2}" in
 	_parallel*)
         rm ${RESULT_DIR}*${2}* 2>/dev/null
 		NEW_RESULT_FILE=${RESULT_FILE}${2}
-		echo "${HEADER}" > ${NEW_RESULT_FILE}
+		if [ ! -s "${NEW_RESULT_FILE}" ]; then
+			echo "${HEADER}" > ${NEW_RESULT_FILE}
+		fi
 		netperf -H ${SERVER_IP} -l 10 | tail -n 1 >> ${NEW_RESULT_FILE}
 	        mv ${NEW_RESULT_FILE} ${NEW_RESULT_FILE}.txt
 		;;
